requirements.txt
# Required Python libraries for Shoplifting Detection System
numpy
opencv-python
pytorch
torchvision
yaml
pandas
ultralytics